==> analyses/survival-analysis/02-summary-statics-TP53.Rmd <==
---
title: "tp53_score_plot_summary"
author: "Zhuangzhuang Geng"
date: "2023-09-19"
output: html_document
---

### load libraries

```{r libraries}
suppressWarnings({
  library(tidyverse)
  library(survival)
  library(ggplot2)
  library(patchwork)
  library(ggpubr)
  library(ggalluvial)
})
```

remove warning
```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

### set directories adn source functions for plotting 

```{r}
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
analyses_dir <- file.path(root_dir, "analyses", "survival-analysis")
plots_dir <- file.path(analyses_dir, "plots")

if(!dir.exists(plots_dir)){
  dir.create(plots_dir)
}

## source this script for plotting
source(file.path(analyses_dir, "utils", "theme_for_plot.R"))
```

### load files and modify the hist and tp53 score

```{r}
hist <- readr::read_tsv(file.path(root_dir, "data", "v1", "Hope-GBM-histologies.tsv"))
tp53_score <- readr::read_tsv(file.path(root_dir, "analyses", "tp53_nf1_score", "results", "tp53_altered_status.tsv"))
ancestry <- readr::read_tsv(file.path(root_dir, "analyses", "survival-analysis", "input", "D3b_Hope_Ad_Hoc_DEI.somalier-ancestry.tsv"))

## add sample_id to ancestry
ancestry <- ancestry %>% 
  dplyr::rename("Kids_First_Biospecimen_ID" = "#sample_id") %>% 
  left_join(hist[, c("Kids_First_Biospecimen_ID", "sample_id")])
  
hist <- hist %>% 
  ## molecular subtype as NA
  filter(!sample_id %in% c("7316-1723", "7316-1746", "7316-194", "7316-212")) %>%
  mutate(broad_mol = case_when(grepl(",", molecular_subtype) ~ substr(molecular_subtype, 1, regexpr(",", molecular_subtype) - 1), 
                               TRUE ~ molecular_subtype), 
         mol_subtype_without_tp53 = case_when(grepl("IHG", molecular_subtype) ~ "IHG", 
                               TRUE ~ gsub(", TP53", "", molecular_subtype)))


tp53_score <- tp53_score %>%
  mutate(binned_tp53 = case_when(tp53_score >= 0 & tp53_score < 0.25 ~ "1 quantile", 
                                 tp53_score >= 0.25 & tp53_score < 0.5 ~ "2 quantile", 
                                 tp53_score >= 0.5 & tp53_score < 0.75 ~ "3 quantile", 
                                 tp53_score >= 0.75 & tp53_score <= 1 ~ "4 quantile", 
                                 TRUE ~ "no tp53 score"))


```

### add tp53 status

```{r}

tp53_score <- tp53_score %>% 
  mutate(SNV_alt = case_when(SNV_indel_counts == 1 ~ "SNV", TRUE ~ "None"),
         CNV_alt = case_when(!is.na(CNV_loss_evidence) ~ "CNV loss", TRUE ~ "None"), 
         Fusion_alt = case_when(!is.na(Fusion_evidence) ~ "Fusion", TRUE ~ "None"), 
         SV_alt = case_when(!is.na(SV_type) ~ "SV", TRUE ~ "None")) %>% 
  mutate(status = paste(SNV_alt, CNV_alt, Fusion_alt, SV_alt, sep = ","), 
         status = gsub(",None", "", status), 
         status = gsub("None,", "", status))
                                     

```


### extract HGG samples and combine histologies, tp53_score table, molecular_subtypes together

```{r}
## extract HGG samples and combine histologies, tp53_score table, molecular_subtypes together
HGG_survival <- hist %>%
  left_join(tp53_score, by = "sample_id") %>%
  ## add predicted_ancestry
  left_join(ancestry[, c("sample_id", "predicted_ancestry")]) %>%
  ## filter out normal samples as they do not have a molecular subtype
  filter(sample_type != "Normal") %>%
  ## remove CPTAC samples 
  filter(!grepl("^C3", sample_id)) %>%
  ## generate a column for age_at_diagnosis_years
  mutate(age_at_diagnosis_year = as.integer(age_at_diagnosis_days/365.25)) %>%
  ## change NA in molecular_subtype and reported gender to "not reported"
  mutate(molecular_subtype = replace_na(molecular_subtype, "Not reported"), 
         reported_gender = replace_na(reported_gender, "not reported")) %>%
  ## change NA in predicted_ancestry to "not reported
  mutate(predicted_ancestry = replace_na(predicted_ancestry, "Not reported")) %>%
  select(sample_id, tp53_score, tp53_altered, 
         OS_days, OS_status, EFS_days, EFS_event_type, 
         reported_gender, race, age_at_diagnosis_days, 
         age_at_diagnosis_year, molecular_subtype, 
         binned_tp53, broad_mol, mol_subtype_without_tp53, status, 
         predicted_ancestry) %>% 
  mutate(binned_age = case_when(age_at_diagnosis_year >= 0 & 
                                  age_at_diagnosis_year <= 15 ~ "[0,15]", 
                                age_at_diagnosis_year > 15 & 
                                  age_at_diagnosis_year <= 26 ~ "(15,26]", 
                                age_at_diagnosis_year > 26 & 
                                  age_at_diagnosis_year <= 40 ~ "(26,40]",
                                age_at_diagnosis_year > 40 & 
                                  age_at_diagnosis_year <= 60 ~ "(40,62]", 
                                TRUE ~ "not reported")) %>%
  distinct() %>% 
  write_tsv(file.path(file.path(analyses_dir, "input", "HGG_survival_tp53.tsv")))

## GBM samples
GBM_df <- hist %>%
    filter(grepl("^C3", sample_id)) %>%
    left_join(tp53_score, by = "sample_id") %>%
    ## transfer CPTAC samples age to column age_at_diagnosis_days
    mutate(age_at_diagnosis_days = (HARMONY_age * 365.25),
           age_at_diagnosis_year = HARMONY_age, 
           reported_gender = HARMONY_Gender, 
           molecular_subtype = "GBM", 
           binned_age = HARMONY_age_class_derived, 
           broad_mol = "GBM", 
           mol_subtype_without_tp53 = "GBM") %>% 
    select(sample_id, tp53_score, tp53_altered, 
         OS_days, OS_status, EFS_days, EFS_event_type, 
         reported_gender, race, age_at_diagnosis_days, age_at_diagnosis_year,
         molecular_subtype, binned_age, binned_tp53, broad_mol, mol_subtype_without_tp53, status) %>% 
    mutate(binned_age = case_when(age_at_diagnosis_year > 62 ~ ">62", TRUE ~ binned_age)) %>%
    distinct() %>% 
    write_tsv(file.path(file.path(analyses_dir, "input", "GBM_survival_tp53.tsv")))

combined <- bind_rows(GBM_df, HGG_survival) %>% 
    mutate(predicted_ancestry = replace_na(predicted_ancestry, "Not reported"))

```

### test if there is a correlation between age and tp53 score

```{r}
## function to print the formula of linear regression 
lm_eqn <- function(df, x, y){
    m <- lm(y ~ x, df);
    eq <- substitute(italic(y) == a + b %.% italic(x)*","~~italic(r)^2~"="~r2, 
         list(a = format(unname(coef(m)[1]), digits = 2),
              b = format(unname(coef(m)[2]), digits = 2),
             r2 = format(summary(m)$r.squared, digits = 3)))
    as.character(as.expression(eq));
}

ggplot(combined, aes(x = age_at_diagnosis_days, y = tp53_score)) + 
  geom_smooth(method = "lm") + 
  geom_point() + 
  geom_text(y = 0.05, x = 20000, 
            label = lm_eqn(combined, x= combined$age_at_diagnosis_days, y = combined$tp53_score), 
            parse = TRUE) + 
  ylab("TP53 score") + 
  xlab("Age at diagnosis days") + 
  theme_Publication()
```

### set a seed for reproducibility of jitter

```{r}

set.seed(2023)
```

### The distribution of TP53 score between age range and tp53 score 

```{r}
group_summary <- combined %>%
  group_by(binned_age) %>%
  summarize(count = n())

# Filter the groups with at least 3 samples
combined_filtered <- combined %>%
  filter(binned_age %in% group_summary$binned_age[group_summary$count >= 3])

ggplot(combined_filtered, aes(x = factor(binned_age, levels = c("[0,15]", "(15,26]", "(26,40]", "(40,62]", ">62", "not reported")), y = tp53_score)) + 
  geom_boxplot() +
  geom_jitter(aes(color = broad_mol), width = 0.3, alpha = 0.5) + 
  stat_compare_means(method = "wilcox", comparisons = list(c("[0,15]", "(15,26]"), c("[0,15]", "(26,40]"), c("[0,15]", "(40,62]"), c("[0,15]", ">62"))) + 
  stat_compare_means(method = "kruskal", label.y = 1.5) + 
  xlab("") + 
  ylab("TP53 score") + 
  labs(color = "Histology") + 
  theme_Publication()

```

### save the plot

```{r}
ggsave(file.path(plots_dir, "boxplot_tp53_score_binned_age.pdf"))

```

### plot to show the tp53 alteration status

```{r}
## change the order of tp53 status legend
combined_filtered <- combined_filtered %>% 
  mutate(status = factor(status, levels = c("CNV loss", "SNV", "SNV,CNV loss", "SNV,Fusion", "None")))

ggplot(combined_filtered, aes(x = factor(binned_age, levels = c("[0,15]", "(15,26]", "(26,40]", "(40,62]", ">62", "not reported")), y = tp53_score)) + 
  geom_boxplot() +
  geom_jitter(aes(color = status), width = 0.3, alpha = 0.5) + 
  xlab("") + 
  ylab("TP53 score") + 
  theme_Publication() 
```

### save the plot

```{r}
ggsave(file.path(plots_dir, "boxplot_tp53_score_status.pdf"), width = 7, height = 5)

```


### test tp53 score difference between the reported gender
```{r}

p1 <- ggplot(HGG_survival, aes(x = reported_gender, y = tp53_score)) + 
      geom_boxplot() +
      geom_jitter(aes(color = broad_mol), width = 0.3, alpha = 0.5) +
      scale_color_manual(values = c("DHG" = "gold2", "DMG" = "red4", "HGG" = "cyan4", "IHG" = "hotpink2", "PXA"  ="olivedrab4")) +
      stat_compare_means(comparisons = list(c("Female", "Male"))) + 
      ggtitle(label = "HOPE cohort") + 
      xlab("") +
      ylab("TP53 score") + 
      theme_Publication() + 
      theme(legend.position="none")


p2 <- ggplot(GBM_df, aes(x = reported_gender, y = tp53_score)) + 
      geom_boxplot() +
      geom_jitter(aes(color = broad_mol),width = 0.3, alpha = 0.5) + 
      scale_color_manual(values = c("GBM" = "purple3")) + 
      stat_compare_means(comparisons = list(c("Female", "Male"))) + 
      ggtitle(label = "GBM") + 
      xlab("") + 
      ylab("TP53 score") + 
      theme_Publication() +
      theme(legend.position="none")
     
p3 <- ggplot(combined, aes(x = reported_gender, y = tp53_score)) + 
      geom_boxplot() +
      geom_jitter(aes(color = broad_mol), width = 0.3, alpha = 0.5) + 
      scale_color_manual(values = c("DHG" = "gold2", "DMG" = "red4", "GBM" = "purple3", "HGG" = "cyan4", "IHG" = "hotpink2", "PXA"  ="olivedrab4")) +
      stat_compare_means(comparisons = list(c("Female", "Male"))) + 
      ggtitle(label = "All samples") + 
      xlab("") +      
      ylab("TP53 score") + 
      labs(color = "Histology") + 
      theme_Publication()

p1 + p2 + p3

```

### save the plot

```{r}
ggsave(file.path(plots_dir, "boxplot_tp53_score_gender.pdf"), height = 6, width = 12)

```


### The distribution of TP53 score among different races 

```{r}
group_summary <- HGG_survival %>%
  group_by(race) %>%
  summarize(count = n())

# Filter the groups with at least 3 samples
HGG_filtered <- HGG_survival %>%
  filter(race %in% group_summary$race[group_summary$count >= 3])

p1 <- HGG_filtered %>% 
  mutate(race = fct_reorder(race, -tp53_score, .fun='median' )) %>%
  ggplot(aes(x = race, y = tp53_score)) + 
  geom_boxplot() +
  geom_jitter(width = 0.3, alpha = 0.5) + 
  xlab("") +
  stat_compare_means(method = "wilcox", comparisons = list(c("Asian", "Reported Unknown"), c("Asian", "Black or African American"), c("Asian", "White"), c("Asian", "Other"))) +
  stat_compare_means(method = "kruskal", label.y = 1.6) + 
  theme_Publication() + 
  ylab("TP53 score") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

p1
```

### save the plot

```{r}
ggsave(file.path(plots_dir, "boxplot_tp53_score_race.pdf"))

```

### The distribution of TP53 score among predicted ancestry

```{r}
group_summary <- HGG_survival %>%
  group_by(predicted_ancestry) %>%
  summarize(count = n())

# Filter the groups with at least 3 samples
HGG_filtered <- HGG_survival %>%
  filter(predicted_ancestry %in% group_summary$predicted_ancestry[group_summary$count >= 3]) %>% 
  filter(predicted_ancestry != "Not reported")

p1 <- HGG_filtered %>% 
  mutate(predicted_ancestry = fct_reorder(predicted_ancestry, -tp53_score, .fun='median' )) %>%
  ggplot(aes(x = predicted_ancestry, y = tp53_score)) + 
  geom_boxplot() +
  geom_jitter(width = 0.3, alpha = 0.5) + 
  xlab("") +
  stat_compare_means(method = "wilcox", comparisons = list(c("EAS", "AFR"), c("EAS", "EUR"), c("EAS", "AMR"))) +
  stat_compare_means(method = "kruskal", label.y = 1.6) + 
  theme_Publication() + 
  ylab("TP53 score") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

p1
```

### save the plot

```{r}
ggsave(file.path(plots_dir, "boxplot_tp53_score_ancestry.pdf"))

```

## alluvial plot for race and predicted-ancestry

```{r}
HGG_race <- as.data.frame(table(HGG_survival$race, HGG_survival$predicted_ancestry)) %>% 
  dplyr::rename(race = Var1, 
                predicted_ancestry = Var2) %>%
  to_lodes_form(axes = 1:2) %>% 
  dplyr::rename(Group = stratum) %>% 
  filter(Group != "Not reported") %>% 
  mutate(Group = factor(Group, levels = c("Asian", "Black or African American", "Native Hawaiian or Other Pacific Islander", "Other", "Reported Unknown", "White", "EAS", "SAS", "AFR", "EUR", "AMR"))) %>% 
  mutate(race = case_when(Group %in% c("Asian", "Black or African American", "Native Hawaiian or Other Pacific Islander", "Other", "Reported Unknown", "White") ~ Group, 
                          TRUE ~ NA), 
         predicted_ancestry = case_when(Group %in% c("EAS", "SAS", "AFR", "EUR", "AMR") ~ Group, 
                                        TRUE ~ NA))


p1 <- ggplot(HGG_race, aes(y = Freq, stratum = Group, alluvium = alluvium, x = x, fill = Group)) + 
  geom_alluvium(show.legend = F) + 
  geom_stratum(show.legend = F) +
  scale_fill_manual(values = c("Asian" = "dodgerblue3", "SAS" = "dodgerblue", "EAS" = "dodgerblue4", 
                                "White" = "gold3", "EUR" = "gold", 
                                "Black or African American" = "brown", "AFR" = "brown3", 
                                "Native Hawaiian or Other Pacific Islander" = "seagreen", "AMR" = "seagreen3", 
                                "Reported Unknown" = "black", "Other" = "grey")) + 
  xlab("") + 
  ylab("Number of Patients") +
  scale_x_discrete(labels = c("reported race", "predicted ancestry")) + 
  theme_Publication()

race_df <- data.frame(race = c("Asian", "Black or African American", "Native Hawaiian or Other Pacific Islander", "White", "Other", "Reported Unknown"), 
                     value = 1)
ancestry_df <- data.frame(predicted_ancestry = c("EAS", "SAS", "AFR", "EUR", "AMR"), 
                          value = 1)

lgd1 <- ggplot(race_df, aes(x = value, y = factor(race, levels = c("Reported Unknown","Other","White","Native Hawaiian or Other Pacific Islander","Black or African American","Asian")), 
                            fill = race)) + 
  geom_tile(show.legend = F) + 
  scale_fill_manual(values = c("Asian" = "dodgerblue3",  
                                "White" = "gold3", 
                                "Black or African American" = "brown", 
                                "Native Hawaiian or Other Pacific Islander" = "seagreen", 
                                "Reported Unknown" = "black", "Other" = "grey")) +
  labs(title = "reported race") + 
  scale_y_discrete(position = "right", expand = c(0, 0)) + 
  scale_x_discrete(expand = c(0, 0)) + 
  coord_fixed(ratio=1) +
  theme(legend.position = "none",
        plot.title = element_text(size=8, face = "bold", colour = "black"),
        panel.grid = element_blank(),
        axis.title = element_blank(),
        axis.text.x = element_blank(),
        axis.ticks = element_blank(),
        axis.text = element_text(size=8, face = "bold", colour = "black"),
        panel.background = element_blank()) 

lgd2 <- ggplot(ancestry_df, aes(x = value, y = factor(predicted_ancestry, 
                                                      levels = c("SAS","EAS","EUR","AMR","AFR")), 
                                fill = predicted_ancestry)) + 
  geom_tile(show.legend = F) + 
  #xlim() + 
  scale_fill_manual(values = c("SAS" = "dodgerblue", 
                               "EAS" = "dodgerblue4",
                               "EUR" = "gold", 
                               "AFR" = "brown3", 
                               "AMR" = "seagreen3")) +
  labs(title = "predicted ancestry") + 
  scale_y_discrete(position = "right", expand = c(0, 0)) + 
  scale_x_discrete(expand = c(0, 0)) + 
  coord_fixed(ratio=1) + 
  theme(legend.position = "none",
        plot.title = element_text(size=8, face = "bold", colour = "black"),
        panel.grid = element_blank(),
        axis.title = element_blank(),
        axis.text.x = element_blank(),
        axis.ticks = element_blank(),
        axis.text = element_text(size=8, face = "bold", colour = "black"),
        panel.background = element_blank()) 
  
  

p1 + lgd1/lgd2
```

## save the plot

```{r}
ggsave(file.path(plots_dir, "alluvial_race_ancestry.pdf"), height = 6, width = 10)

```

### tp53 score distribution about different molecular subtypes 

```{r}
combined_filtered <- combined %>% 
  filter(!is.na(tp53_score)) 

p1 <- combined_filtered %>% 
  mutate(broad_mol = fct_reorder(broad_mol, -tp53_score, .fun='median')) %>%
  ggplot(aes(x = broad_mol, y = tp53_score)) + 
  geom_boxplot() +
  geom_jitter(aes(color = tp53_altered), width = 0.3, alpha = 0.5) +
  xlab("") + 
  ylab("TP53 score") + 
  labs(color = "TP53 status") + 
  theme_Publication()

p2 <- combined_filtered %>% 
  mutate(mol_subtype_without_tp53 = fct_reorder(mol_subtype_without_tp53, -tp53_score, .fun='median')) %>%
  ggplot(aes(x = mol_subtype_without_tp53, y = tp53_score)) + 
  geom_boxplot() +
  geom_jitter(aes(color = tp53_altered), width = 0.3, alpha = 0.5) +
  xlab("") + 
  ylab("TP53 score") + 
  theme_Publication() + 
  labs(color = "TP53 status") + 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))


p1 / p2


```

### save the plot

```{r}
ggsave(file.path(plots_dir, "boxplot_tp53_score_subtype.pdf"), width = 7, height = 8)

```

### session info
```{r}

sessionInfo()
```
